iscsi param cleanup
Remove inconsistencies in param structures and index usage.  This
seems like the cleanest way to keep the three separate lists of 19
parameters in sync.  The three spots are:  lookup of special params
by index #defines during login, session defaults, and target
defaults.  This fixes one real (although oddball) bug and several
potential bugs.

The real bug left the recv dlength size at 8k in spite of both
initiator and target wanting 256k: text_add_key thought 256k was
default and did not bother to send it.  The OFMarker code was errant
too, but it is unlikely that that was ever a problem for someone.

Target parameter defaults are not changed.  Except for the case
where you asked for a dlength of 256k using an explicit iscsiadm
target update command.

Further cleanups to iscsi_if.h would be possible.  Most of that is
legacy code from the target-in-kernel days.

Signed-off-by: Pete Wyckoff <pw@osc.edu>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>