iscsi: no spin while waiting commands to complete
A missing call to tgt_event_modify(fd, EPOLLIN) is added.  Without this, an
asynchronous backing store would cause tgtd to spin in epoll, always
reporting EPOLLOUT on the socket but having nothing to write.  When the
thread completes the command, the bs will cause cmd_done, turning EPOLLOUT
back on.

No change for synchronous backing stores, as epoll does not get called
until the command completes.

It would also spin briefly waiting for the rest of the unsolicited packets,
in both sync and async cases.

Signed-off-by: Pete Wyckoff <pw@osc.edu>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>