iscsi: add transport mxdsl
The RDMA transport chooses the size for data-in and data-out
transfers, as these operations are not visible to the initiator.
MaxXmitDataSegmentLength applies only to control-type PDUs, not to
RDMA data transfers.  Add a per-connection parameter to hold the data
transfer size chosen by the transport, with the proper default for
TCP.

Signed-off-by: Pete Wyckoff <pw@osc.edu>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>