add profile support for mmc
This adds profile support and state to the mmc emulation.  Implement
new, mmc specific, lu-online and lu-offline methods so that when a mmc
device is brought online, we can decide between which of three
different profiles should be used :

PROFILE_NO_PROFILE : if there is no backing store file at all, in
which case we pretend to be a drive with no media inserted.

PROFILE_DVD_PLUS_R : if the backing store file is "empty" (file size
== 0) in which case we pretend we have a blank writeable DVD+R disk in
the tray.

PROFILE_DVD_ROM : if the backing store is non-empty, in which case we
pretend we have a ROM disk in the tray.

During lun initialization, we force the use of the "rdwr"
backingstore.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>