add infrastructure to handle commands taking service actions
implement Maintenance / ReportAllSupportedOpcodes as an example

Expand device_type_operations to also include a list of possible service
actions for this cdb. We need to know how many/what service actions
a certain cdb can take in order to be able to report
SupportedOpcodes properly. (MaintIn/service action 0x0c)

Implement MaintenanceIn and implement service action 0x0c for this command.
the provided infrastructure should make it easier to add other MaintIn
service actions, of which there are quite a few.

there are many other opcodes that also take service actions and which
could build on this framework:
0xa4 MaintOut
0x7f: variable length cdb
0x5e/0x5f persistent reservarion in/out
0xab
0x9e
0x9f

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>