fix MODE_SENSE memory corruption bug
This fixes the bug that MODE_SENSE with small allocation length leads
to memory corruption (due to memcpy overwrite).

Thanks to Ronnie Sahlberg for finding this bug, sending a patch, and
testing this patch:

http://www.mail-archive.com/stgt-devel@lists.berlios.de/msg00523.html

This fixed the bug in a slightly different way (build_mode_page takes
only one pointer where to write the data).

Thanks to Konrad Bechler for helping me to figure out the bug that
tgtd crashes when a Windows XP box formats disk, and testing this
patch:

http://www.mail-archive.com/stgt-devel@lists.berlios.de/msg00553.html

Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>