fcoe: handle fcoe startup errors properly
When we start up fcoe on an interface, check that we actually have an
interface name passed to us and print an error and return -ENODEV
othervise we will SEGV trying to memcpy() from a null pointer 1 line
further down.

When we initialize fcoe from the fcoe module, return proper
success/failure back to tgtd so that it can abort/fail if the low
level driver failed to initialize.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>