enlarge SCSI_SN_LEN
Define a larger SCSI_SN_LEN in order to export the real device's
serial number.

Today, SCSI_SN_LEN is defined as 8 chars only.  I want to define a
larger SCSI_SN_LEN in order to export the real device's serial number.
This is required for using dm-multipath.  When several targets export
the same device to the initiator we will be able to use dm-multipath
on the initiator.

Signed-off-by: Doron Shoham <dorons@voltaire.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>