fcoe: use new header by default
The intel stack which probably most people are going to test with
uses the new header format. We do not have a way to config old or
new like is done with scst fcoe target, and I do not really care to
add much code for that anyways so this just sets the default to the
new format.

The next patch will remove all the old format code.

Signed-off-by: Mike Christie <michaelc@cs.wisc.edu>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>