tgt-admin: add support to change the state of targets
This adds a possibility to change the state (offline or ready) of
targets. It is possible to offline/ready all targets, target specified
with its name (i.e. iqn.2008-08.com.example:some.target) or tid.

This is an example usage help for the --offline option (--ready works
in the same way):

# ./tgt-admin --offline help
"./tgt-admin --offline <value>" - offline all or selected targets.

Example usage:
        --offline help  - display this help
        --offline ALL   - offline all targets
        --offline tid=4 - offline target 4 (target with tid 4)
        --offline iqn.2008-08.com.example:some.target - offline this target

Signed-off-by: Tomasz Chmielewski <mangoo@wpkg.org>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>