tgt-admin: improved delete option for removing targets
This patch adds a better --delete option which is used for removing
targets.

It works similarly to --offline and --ready options:

--delete <value>          delete all or selected targets
                          The target will be deleted only if it's not used
                          (no initiator is connected to it).
                          If you want to delete targets which are in use,
                          you have to add "--force" flag

Example usage:
      --delete help           - display this help
      --delete ALL            - delete all targets
      --delete tid=4          - delete target 4 (target with tid 4)
      --delete iqn.2008-08.com.example:some.target - delete this target

Basically, it will remove a target if it's not in use; if you still
want to remove the target even though it's in use, you have to add
--force.

Because of this, I rename the previous --force into --ignore-errors
(previously, --force was used to continue execution even if tgtadm
exited with non-zero code).

Also, "-d" (previously, a short for --delete) is gone and is not an
option anymore. This is because --delete is potentially a dangerous
operation so some extra typing could be justified here. Moreover, lots
of programs use "-d" to enable debugging, so we don't want to use it.

Currently, --delete is very quiet; if you want some output, please add
a -v/--verbose option.

Signed-off-by: Tomasz Chmielewski <mangoo@wpkg.org>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>