tgt-admin: add support to update targets
This patch adds proper updating of targets to tgt-admin.

You have to add --force/-f option to update targets which are in use.

Updating consists of two things:
- target is deleted
- target is configured, basing on its config options

Updating a target which is no longer in a config file will result in
the target's deletion.

The patch also changes the behaviour of --execute slightly - it now
tries to delete the targets which are not in the config file; if the
target is in use, it won't be touched (unless --force is used); if the
target is not in use, it will be deleted.

On my TO DO list are still:
- don't assume cid is 0 when removing targets (check for a proper cid)
- some code cleanups
- shutting down tgtd (only if there are no targets)
- adding some more options (disabling write-cache etc.)

Signed-off-by: Tomasz Chmielewski <mangoo@wpkg.org>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>