add SCSI Generic driver (sg) based backing store
Backing store based on SCSI Generic driver (sg), v.3.
allows to route scsi commands to the native scsi devices
unchanged. It uses Direct I/O (dio) to avoid memcopy
to the buffer cache.

Using sg allows asynchronous execution, commands are
written to a sg-backed char device, completions are
read from the same device, which can be polled
asynchronously.

This bs provides significant performance improvement when
working with native scsi devices. In a setup, where
the scsi devices are exported by a tgt with bs_null,
and both links (from initiator to target and from the
target to the "backing-store" target) are iSER/IB
sustained bandwidth of 1450 MB/s for READ and
1350 MB/s for WRITE is achieved. This to be compared to
700-800 MB/s when running with bs_rdwr in the same setup.
Some improvements are seen with IOPS as well:
60 kIOPS for READ, 38 kIOPS for WRITE
(compared to 31/35KIOPS with bs_rdwr).

Signed-off-by: Alexander Nezhinsky <nezhinsky@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>