iscsi: fix session reinstatement ref-count mess-up
login_security_done calls conn_close() in case of session
reinstatement. if conn_close() was already called against a
connection, it messes up the connection ref-counting. conn_close()
must be called only once for a connection.

This patch introduces the closed to struct iscsi_connection to avoid
multiple calls of conn_close but it should be a new state.

Thanks to Tomasz Chmielewski <mangoo@wpkg.org> for helping me fix the
bugs.

Reported-by: Tomasz Chmielewski <mangoo@wpkg.org>
Tested-by: Tomasz Chmielewski <mangoo@wpkg.org>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>