tgt-admin: allow --ignore-errors to be used in the config file
Allow --ignore-errors to be used in the config file.

Rationale: if any of the devices used by the target is inaccessible
(i.e., because it's now a secondary node in DRBD), tgt-admin will not
process all remaining targets, but exit with error somewhere in the middle.
As a result, any (re)connecting initiators may not find the target
they expect, even if there wouldn't be any errors when configuring
that target.
This is especially painful if the whole target machine is being restarted
and tgtd startup/configuration happens automatically, in the background.

Signed-off-by: Tomasz Chmielewski <mangoo@wpkg.org>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>