revert "force ssc to use bs_ssc"
This reverts commit d7a21bac046516484efbdf7f0c01d5e747482ebe

On the error path, the patch tries to free an invalid memory and then
tgtd crashes. We can add a trick to avoid this but this patch is a
hack. We need to implementation a new way to represent what backing
store types each device type can handle. So I add a trick to tgtadm to
force ssc to use bs_ssc for now.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>