spc: change vpd page 0x83 scheme
1. doesn't report protocol id anymore (PIV set to zero)
2. reports LUN association (zero) instead of target port association
3. reports T10 identifier type instead of vendor specific
4. uses the T10 IET assigned name in the identifier

The first change is necessary because we don't report protocol id
properly (we always uses iSCSI). Let's just drop PIV bit (means that
the protocol id is not supported) as the majority of SCSI target
devices do.

The second, third and fourth changes are for VMware: VMware needs LUN
association; The vendor specific identifier type is not among the list
of VMware supported id types and T10 is.

Thanks to Or Gerlitz <ogerlitz@voltaire.com> for pointing out this
problem and providing a first version of the fix.

The further information is available at:

http://lists.wpkg.org/pipermail/stgt/2009-February/002652.html

Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>