unlink a file for AF_LOCAL
The commit 49e93cd2f02bbbbaa18ca3498ba4580338ec5045 changed the
management socket to use a file instead of an abstract namespace,
which only Linux supports. So we need to unlink a file for
AF_LOCAL. We don't care if the file exists or doesn't so we don't
checked the returned value of unlink.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>