iscsi: fix cmd sequence numbering for login phase of discovery sessions
For discovery sessions the Exp- and the MaxCmdSNs are not set appropriately
during login phase. This leads to e.g. the OpenSolaris initiator not being able to
successfully perform a SendTargets discovery.
This patch also increases the CmdSN window size to avoid an issue previously observed
with the AIX initiator vs. IET (not tested - cf. IET svn rev. 22).

Signed-off-by: Arne Redlich <arne.redlich@googlemail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>