add documentation about the tgtd daemon in XSLT format
If xsltproc is available, use this tool to compile this XSLT file into
both a manpage and a html page for tgtd.

Add a directory doc/htmlpages to store html versions of manpages.
These could go to /usr/share/doc/tgt ?

Check in both generated file, man and html, so that they are still
available on systems where xsltproc is not available to build them.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>