doc: update the tgtadm manpage and add two new sections
LUN PARAMETERS that describe parameters that apply to a LUN

SMC SPECIFIC LUN PARAMETERS that describes parameters specific
to the media changed device type.

This now replicates nost of the information from
README.lu_configuration in the more accessible tgtadm manpage.

As an example on how to configure a SMC device at the end of the
SMC section I copied the "how to create a dvd jukebox" example
from README.mmc

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>