tgtimg: sbc support
Use tgtimg to create DISK images. This is a trivial reimplementation
of dd if=/dev/zero bs=1M ... but offers us to use one command for any
type of media.

Update the manpage for TGTIMG to describe how to use TGTIMG
to create DISK media.

Change the example in README.iser to use TGTIMG instead of dd to
create the mediafile.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>