Moved check for is_debug from log_debug() function to dprintf() macro
To minimize the overhead of debug printouts during non-debug runs,
the check for "is_debug" is moved from log_debug() function
to dprintf() macro.

Now log_debug is called only if "is_debug" is set.
This spares a function call, parameter passing and the costs of
complex arguments expanding and evaluation.
To minimize the potential cost of log_debug() even more,
the check for is_debug is put under unlikely() statement.

Signed-off-by: Alexander Nezhinsky <alexandern@voltaire.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>