iscsi: fix isns_attr_query() and send_scn_rsp() buffer overflow
print_snc_pdu() doesn't check the length of ISNS_ATTR_ISCSI_NAME so
that a bad pdu with a long iSCSN name could cause a buffer overflow in
isns_attr_query() and send_scn_rsp().

With this patch, the maximum length of iSCSI names that
print_snc_pdu() returns is 223 bytes (as iSCSI RFC defines). The
buffer length in isns_attr_query() and send_scn_rsp() is long
enough.

Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>