iscsi: use a local variable for enabling/disabling entity
Remove the dependence of iscsi_targets_list in isns_target_register() and
isns_target_deregister() to determine whether to enable/disable the
entity. Instead count the number of targets that are registered with the
SNS Server.

This is in preparation of the following patch that registers targets
that are already defined when iSNS is turned On.

Signed-Off-By: Chandra Seetharaman <sekharan@us.ibm.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>