iscsi: remove function isns_deregister()
Function isns_deregister() is called from isns_exit() to deregister the
target from SNS server, and currently isns_exit() is not called.

This function is not necessary as target deregister from iSNS server
happens in isns_target_deregister(),

This patch is in preparation of the following patch to allow
turning iSNS on/Off.

Signed-Off-By: Chandra Seetharaman <sekharan@us.ibm.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>