add a new device parameter "readonly={0|1}"
Added a device parameter that allows a logical unit to be
set to readonly. For DISK devices, this is communicated
back to the initiator by setting the WP bit in the device
specific byte in the mode sense header. It will also fail
any WRITE6/10/12/16 calls with a DATA_PROTECT/ASC_WRITE_PROTECT

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: Mike Anderson <andmike@linux.vnet.ibm.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>