iscsi: release of the current tx task on conn_close
Fixes a bug with user-initiated disconnect that led to failures in
deleting the logical unit and target, during concurrent I/O.

Typically, the error messages were:
> tgtadm: this logical unit is still active
> tgtadm: this target is still active
(as discussed in the thread "back to forced deletion" starting with
http://lists.wpkg.org/pipermail/stgt/2010-September/004095.html)

The current tx task (conn->tx_task) was freed using the basic function
iscsi_free_task(), while all other pending tx tasks are freed differentially,
according to their type.

The current tx task also should be freed differentially, or else the task
hash table of the nexus remains non-empty and the target is declared active.

To achieve this the current tx task is added to the list of pending tx tasks
just before their release, so that it is freed according to the same policy.

Signed-off-by: Alexander Nezhinsky <alexandern@voltaire.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>