new timer-based work scheduling
Re-implementing the time-based work scheduler. This patch implements
a timer-based scheme.

Work items are scheduled from various application contexts, and put on
a queue. The current time from gettimeofday() is used instead of
a jiffie-based mechanism used previosuly. The work item is stamped with
the expiration time, obtained as the current time plus the timeout period.

A global timer (one per process) is registered and fired periodically,
few times a second. When the timer signal is handled, a word is written
to a dedicated pipe, whose fd is registered with the event loop.

The event handler reads from the pipe, and examines the inactive work
queue. All items that have expired (the immediate current time ise used
for comparison, not the timer or pipe event handling times) are moved
to the active list and processed one after another. The items that
might have accumulated and expired in the meantime are not handled
and are postponed until the next timer event.

The new scheme handles the timer event through the standard event loop's
file descritor registration and polling. This removes the previosly used
call to schedule() function which had to update jiffies.

There is still no guarantee about the exact work execution time.
This scheme is suitable for tasks with seconds-scale resolution, so that
firing the timer few times per second provides a satisfactory accuracy.
The assumption is that all event handlers take small periods of time.
If not, in theory, the timer handlers may be delayed indefinitely,
but then the entire event loop processing gets stalled.

The immediate clients of this mechansim is the iSNS code and the new
iSER code to be based on it instead of the previously proposed
custom timer.

Signed-off-by: Alexander Nezhinsky <alexandern@voltaire.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>