iscsi: handle tx properly when iser connection is closing
Removed ignoring tx when connection is in STATE_EXIT, which means that
we are going to close the connection because of a logical error
(presumably in login), but still have to send something.

Alternatively, when we close a connection asynchronously, iser_conn_close()
gets called, and sets the state to STATE_CLOSE. After that we really must
ignore any scheduled tx.

Related to that, in iser_login_tx_complete() we don't need to upgrade
STATE_EXIT to STATE_CLOSE, because STATE_CLOSE is supposed to be set only
from iser_conn_close().

Signed-off-by: Alexander Nezhinsky <alexandern@voltaire.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>