fixes in call_program (closing pipe fds, waitpid interrupted by a signal)
Some fixes in call_program():

1) if fork() fails, previously open pipe file descriptors should be
closed.

2) the redirect script is invoked from a fork'ed process so tgt has to
wait for its child. Currently this is done synchronously using
waitpid().  Sometimes the system call is interrupted by a signal (most
probably by a SIGCHLD signal from the very process that is being
waited upon).  To overcome this, waitpid() is called repeatedly while
the error code is "Interrupted by a signal". Other reasons for failure
are non-existing process or invalid arguments, so in these cases we
shall just give up.

Signed-off-by: Alexander Nezhinsky <alexandern@voltaire.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>