redirect: protect again tgtd process hang as of cluster software hang
If the child process spawned to run the redirect callback script hangs, e.g
as of load/bug in the application which this script is dealing with, tgt
can hang forever. Protect against that by selecting the fd from which tgt
is expected to read, for up to 100ms, if the timeout expires, tgt terminates
the child process and fail the initiator login attempt. While not being the
ultimate solution, of using tgt_event_add et al. and making the redirect
code fully asynchronous, this patch adds protection and makes things much
better in that respect.

Signed-off-by: Alexander Nezhinsky <alexandern@mellanox.com>
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>