mmc: Properly handle READ TOC command
CD tracks starts at one, not zero.

This bug caused /lib/udev/cdrom_id to fail to identify the disc
contents and as a result, the desktop ignores the disc. With this
bug-fix, this no longer happens.

Also fix the TOC returned so it conforms with 6.25.3.2.4 of MMC-6 (the
returned length was wrong).

https://bugzilla.redhat.com/show_bug.cgi?id=690181

Signed-off-by: David Zeuthen <davidz@redhat.com>
Signed-off-by: Andy Grover <agrover@redhat.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>