mmc: READ TOC track==0 is valid
READ TOC  track==xyz means : return all tracks starting at number xyz.
In case of single session disks, the only track that exists is track 1
in which case we still need to return the information for track 1
regardless of whether the initiator asks for track 0 or track 1.

i.e.  READ TOC track==0 is still a valid command even if track 0 does not exist

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>