spc: add minimal support for SEND-DIAGNOSTICS
This implements the minimum required part of SEND-DIAGNOSTICS
(self-test).

Implement the minimal required part of this opcode. This opcode is
mandatory for sbc and ssc devices but only the SEFT-TEST varient is
required.  Return status good for the default self-test. For all
other varients return ILLEGAL-REQUEST/INVALID_FIELD_IN_CDB

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>