sbc: implement VERIFY 10/12/16 and check the arguments
TGTD does not support formatting protection information so make
TGTD fail a VERIFY command requesting vprotect != 0 with a proper
check condition.

See SBC 5.22 VERIFY 10 COMMAND,  tables 65 and 66
If the media is not formatted with protection information (as in TGTD)
any value of vprotect other than 000b is an error condition.

If BYTCHK==1 we must also verify the DATA-OUT buffer with the content
of the media. Add a check that the data matches and return sense
key:MISMATCH and the proper ASCQ if a mismatch is found.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>