Allow creating SBC devices without any media loaded
When creating a device that is supposed to have a backing file
but no path was provided. Treat this as creating a removable device
without any media loaded.
Set removable=1 and make sure online=0 to indicate this is a
removable device and that all media access commands will fail with
check-condition/not ready/medium not present.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>