add an example on how to create a SBC Jukebox
The jukebox creates a dedicated target with one SBC device (removable
media) and one media changer with some storage element slots attached.
The jukebox script can be used to take "snapshots" of a different file
(a production LUN exported via a different target?) and expose these
snapshots through the snapshot target.

This allows you to do things such as
* export a production LUN to an initiator
* from cron, take a snapshot of the production LUN every xyz
* export the snapshots via the snapshot target to a different host to do xyz
* the other host can then pick and chose which snapshot from the production LUN
it wants to access.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>