Correct fsync on write FILEMARKS
commit: 96e83a2e60857cd5356f99c45f0a0a658f600949

missed fsync() on zero filemarks count.

SSC4r01e states:

NOTE 43 - Upon completion of any buffered write operation, the
application client may issue a WRITE FILEMARKS(6) command with the
IMMED bit set to zero and the FILEMARK COUNT field set to zero to
perform a synchronize operation

The IMMED bit is defined as:

a) written to the object buffer if the IMMED bit is set to one; or
b) written to the medium if the IMMED bit is set to zero.

This patch calls fsync() ignoring the IMMED bit value.

Signed-off-by: Mark Harvey<markh794@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>