DPO: Add emulation of DPO bit for READ/WRITE/VERIFY10/12/16
When the DPO bit is set to 1, this is an indication the initiator will
not reference the data again. For this case, call posix_fadvise() and
let the kernel know it may expunge this data from the page cache.

From SBC READ10 :

NOTE 11 - The DPO bit is used to control replacement of logical blocks
in the cache when the application client has information on the future
usage of the logical blocks. If the DPO bit is set to one, the
application client is specifying that the logical blocks accessed by
the command are not likely to be accessed again in the near future and
should not be put in the cache nor retained by the cache. If the DPO
bit is set to zero, the application client is specifying that the
logical blocks accessed by this command are likely to be accessed
again in the near future.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>