tgtimg: Add support for thin-provisioning
A new flag --ting-provisioning is added to the tgtimg command.
When used this flag will create a sparse file without any
allocation guarantee.
It will also verify that FALLOC_FL_PUNCH_HOLE works.

If such files are created, a future patch to tgtd will add
support for the UNMAP command that is used by initiators to
release blocks that are no longer in use.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>