SBC: initial WRITE_SAME16 support. (no unmap support)
Initial support for WRITE_SAME16.
This command is largely the same as WRITE_SAME10.
No support for using the command to perform UNMAP of blocks on thin provisioned LUNs yet.

WRITE_SAME16 only works for the bs_rdwr backend for now. We can add support to bs_aio at a later stage.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>