SBC: Support unmapping blocks using WRITE_SAME/WRITE_SAME16
Add support for WRITE_SAME and WRITE_SAME16 to unmap blocks if the
UNMAP bit is set, and the LUN is set for thin-provisioning.

Log error if the ANCHOR bit is set since we not not yet support
resource-provisioned luns.

Update VPD page 0xb2 to indicate that both write same 10 and write same 16
can be used for unmapping blocks

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>