improve str_to_int macro
There is a potential problem when a user-supplied value overflows the
accepting variable; to cope with this the string is converted to a
64-bit value first then assigned to the variable, then the macro
checks if they represent the same value. Thus the natural type
boundaries are automatically guarded. This means that there is no need
to use type's range limits (e.g. 0 and USHRT_MAX for unsigned
short). Thus the range limit parameters are removed from str_to_int()
-- it's enough for the (most frequent) cases when any value fitting a
certain integer type is admissible. When the min/max values are
relevant they can be passed to one of the new macros str_to_int_gt,
str_to_int_ge, str_to_int_lt, str_to_int_range, requiring that the int
value should be greater than, greater or equal, less than, or within a
range, correspondingly.

Signed-off-by: Alexander Nezhinsky <alexandern@mellanox.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>