bs_aio: use 64bit to read eventfd counter
when using eventfd, each time the fd becomes readable, the number of
io completions can be read; this number is provided as a 64-bit
integer, thus enough space (at least 8 bytes) should be given to
read(). This patch adds a comment which explicitely states the above
requirement. The actual number of completions in bs_aio is limited by
a very small value, thus when reading, a 64-bit variable should be
used, then it can be casted to unsigned int and processed as such.

Signed-off-by: Alexander Nezhinsky <alexandern@mellanox.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>