SBC: READ/WRITE Better check that the LBA is not out of range
If transfer length is zero, which is not an error, we still have to check
that the LBA is within the valid range.

READ10 in the spec says:

The TRANSFER LENGTH field specifies the number of contiguous logical
blocks of data that shall be read and
transferred to the Data-In Buffer, starting with the logical block
specified by the LOGICAL BLOCK ADDRESS field.
A TRANSFER LENGTH field set to zero specifies that no logical blocks
shall be read. This condition shall not be
considered an error. Any other value specifies the number of logical
blocks that shall be read. If the LBA plus
the transfer length exceeds the capacity of the medium, then the
device server shall terminate the command
with CHECK CONDITION status with the sense key set to ILLEGAL REQUEST
and the additional sense code
set to LOGICAL BLOCK ADDRESS OUT OF RANGE.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>